requirements.txt
Django
djangorestframework
djangorestframework-simplejwt
requests
beautifulsoup4
anthropic
python-dotenv
pandas